Rewrote parse_bed without quick:Csv and moved to path_parser. Made the 12-columns parse more transparent
When the bed is 12-columns the function was taking the block information and using them to split the path segment. This is not default behavior in other tools that deal with bed files, for example bedtools. I kept the same functionality, but added a boolean to make it more transparent in the feature. I should open an issue in the future to make it user defined